Style changes to adhere https://github.com/yiisoft/yii2-coding-standards
Extracted MySphinxTestCase and RuntimeIndexFixture into separate files
Fixed newlines
